requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Babel==2.0.0
Flask-Limiter==3.3.0

# Database
SQLAlchemy==2.0.6
alembic==1.10.2

# Security
Werkzeug==2.2.3
itsdangerous==2.1.2
bcrypt==4.0.1

# Image handling
Pillow==9.4.0

# Environment and configuration
python-dotenv==1.0.0
click==8.1.3

# Utilities
pytz==2022.7.1
email-validator==1.3.1
limits==3.5.0

# Development tools
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.2
